refactor(fields): Simplify field components by integrating label handling and enhancing contentlet support
