+++ core-web/libs/edit-content/src/lib/fields/dot-edit-content-category-field/components/dot-category-field/dot-category-field.component.html
@@ -1,0 +1,27 @@
+@if (store.selected().length) {
+    <dot-category-field-chips
+        data-testId="category-chip-list"
+        [categories]="store.selected()"
+        [disabled]="$isDisabled()"
+        (remove)="store.removeRootSelected($event)" />
+}
+
+<div class="dot-category-field__select">
+    <button
+        type="button"
+        (click)="openCategoriesDialog()"
+        [disabled]="store.isDialogOpen() || $isDisabled()"
+        [label]="'edit.content.category-field.show-categories-dialog' | dm"
+        class="p-button-sm p-button-text p-button-secondary"
+        data-testId="show-dialog-btn"
+        pButton></button>
+</div>
+
+@if (store.isDialogOpen()) {
+    @defer (when store.isDialogOpen()) {
+        <dot-category-field-dialog
+            [isVisible]="store.isDialogOpen()"
+            (closedDialog)="store.closeDialog()"
+            data-testId="dialog-placeholder" />
+    }
+}
