+++ core-web/libs/edit-content/src/lib/fields/dot-edit-content-category-field/dot-edit-content-category-field.component.ts
@@ -1,30 +1,20 @@
-import {
-    ChangeDetectionStrategy,
-    Component,
-    computed,
-    effect,
-    inject,
-    Injector,
-    input,
-    OnInit,
-    signal
-} from '@angular/core';
-import { ReactiveFormsModule } from '@angular/forms';
+import { ChangeDetectionStrategy, Component, inject, input } from '@angular/core';
+import { ControlContainer, ReactiveFormsModule } from '@angular/forms';
 
 import { ButtonModule } from 'primeng/button';
 
 import { DotCMSContentlet, DotCMSContentTypeField } from '@dotcms/dotcms-models';
 import { DotMessagePipe } from '@dotcms/ui';
 
-import { DotCategoryFieldChipsComponent } from './components/dot-category-field-chips/dot-category-field-chips.component';
-import { DotCategoryFieldDialogComponent } from './components/dot-category-field-dialog/dot-category-field-dialog.component';
+import { DotCategoryFieldComponent } from './components/dot-category-field/dot-category-field.component';
 import { CategoriesService } from './services/categories.service';
 import { CategoryFieldStore } from './store/content-category-field.store';
 
 import { DotCardFieldContentComponent } from '../dot-card-field/components/dot-card-field-content.component';
 import { DotCardFieldFooterComponent } from '../dot-card-field/components/dot-card-field-footer.component';
+import { DotCardFieldLabelComponent } from '../dot-card-field/components/dot-card-field-label.component';
 import { DotCardFieldComponent } from '../dot-card-field/dot-card-field.component';
-import { BaseFieldComponent } from '../shared/base-field.component';
+import { BaseWrapperField } from '../shared/base-wrapper-field';
 
 /**
  * @class
@@ -40,27 +30,24 @@ import { BaseFieldComponent } from '../shared/base-field.component';
         ReactiveFormsModule,
         ButtonModule,
         DotMessagePipe,
-        DotCategoryFieldChipsComponent,
-        DotCategoryFieldDialogComponent,
         DotCardFieldComponent,
         DotCardFieldContentComponent,
         DotCardFieldFooterComponent,
+        DotCardFieldLabelComponent,
+        DotCategoryFieldComponent,
         DotMessagePipe
     ],
     templateUrl: './dot-edit-content-category-field.component.html',
-    styleUrl: './dot-edit-content-category-field.component.scss',
     changeDetection: ChangeDetectionStrategy.OnPush,
-    host: {
-        '[class.dot-category-field__container--has-categories]': '$hasSelectedCategories()',
-        '[class.dot-category-field__container]': '!$hasSelectedCategories()',
-        '[class.dot-category-field__container--disabled]': '$isDisabled()'
-    },
-    providers: [CategoriesService, CategoryFieldStore]
+    providers: [CategoriesService, CategoryFieldStore],
+    viewProviders: [
+        {
+            provide: ControlContainer,
+            useFactory: () => inject(ControlContainer, { skipSelf: true })
+        }
+    ]
 })
-export class DotEditContentCategoryFieldComponent extends BaseFieldComponent implements OnInit {
-    readonly store = inject(CategoryFieldStore);
-    readonly #injector = inject(Injector);
-
+export class DotEditContentCategoryFieldComponent extends BaseWrapperField {
     /**
      * The `field` variable is of type `DotCMSContentTypeField` and is a required input.
      * @description The variable represents a field of a DotCMS content type and is a required input.
@@ -71,90 +58,4 @@ export class DotEditContentCategoryFieldComponent extends BaseFieldComponent imp
      * @description DotCMSContentlet input representing a DotCMS contentlet.
      */
     $contentlet = input.required<DotCMSContentlet>({ alias: 'contentlet' });
-
-    protected $isDisabled = signal(false);
-
-    /**
-     * The `$hasConfirmedCategories` variable is a computed property that returns a boolean value.
-     *
-     * @returns {Boolean} - True if there are selected categories, false otherwise.
-     */
-    $hasSelectedCategories = computed(() => this.store.selected().length > 0);
-
-    /**
-     * Initialize the component.
-     *
-     * @memberof DotEditContentCategoryFieldComponent
-     */
-    ngOnInit(): void {
-        // Initialize the store with field information only
-        // The contentlet data will come through ControlValueAccessor's writeValue
-        this.store.load({
-            field: this.$field(),
-            contentlet: this.$contentlet()
-        });
-
-        // Effect to sync selected categories with form control
-        effect(
-            () => {
-                const categoryValues = this.store.selected();
-                const inodes = categoryValues?.map((category) => category.inode) ?? [];
-
-                // Notify form control of value change
-                this.onChange(inodes);
-            },
-            {
-                injector: this.#injector
-            }
-        );
-
-        this.statusChanges$.subscribe(() => {
-            this.changeDetectorRef.detectChanges();
-        });
-    }
-
-    /**
-     * Open the categories dialog.
-     *
-     * @memberof DotEditContentCategoryFieldComponent
-     */
-    openCategoriesDialog(): void {
-        if (this.$isDisabled()) {
-            return;
-        }
-
-        this.store.openDialog();
-        this.onTouched();
-    }
-
-    /**
-     * Sets the value in the component when the form control value changes.
-     * This method is called by Angular's forms system.
-     *
-     * @param value - Array of category inode strings
-     */
-    writeValue(value: string[]): void {
-        if (!value) {
-            this.store.setSelectedFromInodes([]);
-
-            return;
-        }
-
-        if (!Array.isArray(value)) {
-            return;
-        }
-
-        // Update store with the new value
-        this.store.setSelectedFromInodes(value);
-    }
-
-    /**
-     * Sets the disabled state of the component.
-     * This method is called by Angular when the form control's disabled state changes.
-     *
-     * @param isDisabled - Whether the component should be disabled
-     */
-    setDisabledState(isDisabled: boolean): void {
-        this.$isDisabled.set(isDisabled);
-    }
 }
