+++ core-web/libs/edit-content/src/lib/fields/dot-edit-content-relationship-field/components/dot-relationship-field/dot-relationship-field.component.scss
@@ -1,0 +1,63 @@
+@use "variables" as *;
+
+::ng-deep {
+    p-table {
+        .dotTable.p-datatable.p-datatable-relationship {
+            .p-datatable-header {
+                background-color: $color-palette-gray-100;
+            }
+            .p-datatable-footer {
+                padding: $spacing-1 0 0 0;
+            }
+            .p-datatable-wrapper {
+                border: 1px solid $color-palette-gray-300;
+                border-radius: $border-radius-md;
+            }
+            .p-paginator {
+                display: none;
+            }
+            .p-datatable-footer {
+                border: 0;
+                padding-left: 0;
+                padding-right: 0;
+                font-weight: lighter;
+            }
+        }
+    }
+}
+
+.add-item-btn {
+    position: absolute;
+    top: 0.4rem;
+    right: 0.4rem;
+}
+
+.relationship-field__table_header {
+    th {
+        font-weight: $font-weight-bold;
+        background-color: $color-palette-gray-100;
+    }
+}
+
+.drag-handle-column {
+    width: 3rem;
+    min-width: 3rem;
+}
+
+.title-column {
+    width: 12rem;
+    min-width: 12rem;
+}
+
+.dynamic-column {
+    min-width: 8rem;
+}
+
+.image-column {
+    width: 6rem;
+    min-width: 6rem;
+}
+
+.actions-column {
+    width: 4rem;
+}
