+++ core-web/libs/edit-content/src/lib/fields/dot-edit-content-text-field/dot-edit-content-text-field.component.html
@@ -1,5 +1,6 @@
 @let field = $field();
 @let showLabel = $showLabel();
+@let dataType = field.dataType;
 
 <dot-card-field [hasError]="hasError">
     @if (showLabel) {
@@ -9,12 +10,12 @@
     }
     <dot-card-field-content>
         <input
-            [id]="variableName"
-            [name]="variableName"
-            [formControlName]="variableName"
+            [id]="field.variable"
+            [name]="field.variable"
+            [formControlName]="field.variable"
             [class.ng-invalid]="hasError"
             [class.ng-dirty]="hasError"
-            [attr.data-testId]="variableName"
+            [attr.data-testId]="field.variable"
             [type]="inputTextOptions[dataType].type"
             [attr.inputmode]="inputTextOptions[dataType].inputMode"
             [attr.step]="inputTextOptions[dataType].step"
