--- core-web/libs/edit-content/src/lib/fields/dot-edit-content-text-field/dot-edit-content-text-field.component.scss
@@ -1,4 +1,0 @@
-:host {
-    height: fit-content;
-    display: block;
-}
